# backend/docs/swagger.yaml
components:
  schemas:
    data:
      properties:
        data:
          $ref: '#/components/schemas/models.UserResponse'
      type: object
...
    handlers.PaymentMethodConfig:
      properties:
        config:
          additionalProperties: {}
          type: object
        is_enabled:
          type: boolean
        method_type:
          type: string
      required:
      - method_type
      type: object
...
    models.Error:
      description: Error response object
      properties:
        code:
          example: "404"
          type: string
        field:
          type: string
        message:
          type: string
      type: object
    models.ErrorResponse:
      description: Generic API error response
      properties:
        code:
          type: integer
        errors:
          items:
            $ref: '#/components/schemas/models.Error'
          type: array
          uniqueItems: false
        message:
          type: string
        meta:
          $ref: '#/components/schemas/models.Meta'
        status:
          type: string
      type: object
...
    models.Meta:
      description: Meta object
      properties:
        timestamp:
          example: "2025-02-12T18:31:40Z"
          type: string
      type: object
...
    models.SuccessResponse:
      allOf:
      - $ref: '#/components/schemas/data'
      description: Generic API response
      properties:
        code:
          example: 200
          type: integer
        data: {}
        message:
          example: Object updated successfully
          type: string
        meta:
          $ref: '#/components/schemas/models.Meta'
        status:
          example: success
          type: string
      type: object
...
    models.UserResponse:
      properties:
        created_at:
          example: "2025-02-09T09:38:25Z"
          format: date-time
          type: string
        email:
          type: string
        last_login:
          example: "2025-02-09T09:38:25Z"
          format: date-time
          type: string
        locale:
          type: string
        name:
          type: string
        profile_picture:
          type: string
        provider:
          type: string
        provider_id:
          type: string
        shops:
          items:
            $ref: '#/components/schemas/models.UserShopResponse'
          type: array
          uniqueItems: false
        user_id:
          description: 'TODO: add sub fields'
          type: string
      type: object
    models.UserShopResponse:
      properties:
        created_at:
          example: "2025-02-09T09:38:25Z"
          format: date-time
          type: string
        custom_domain:
          type: string
        shop_id:
          type: integer
        status:
          type: string
        subdomain:
          type: string
        title:
          type: string
        updated_at:
          example: "2025-02-09T09:38:25Z"
          format: date-time
          type: string
      type: object
...
  securitySchemes:
    OAuth2AccessCode:
      flows:
        authorizationCode:
          authorizationUrl: http://127.0.0.1:8080/oauth2/auth
          tokenUrl: http://127.0.0.1:8080/oauth2/token
      in: header
      type: oauth2
...
info:
  description: This is the Naytife API documentation
  title: Naytife API Docs
  version: "1.0"
openapi: 3.1.0
paths:
...
  /shops/{shop_id}/payment-methods:
    get:
      description: Retrieve all configured payment methods for a specific shop
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                description: Generic API response
                properties:
                  code:
                    example: 200
                    type: integer
                  data: {}
                  message:
                    example: Object updated successfully
                    type: string
                  meta:
                    $ref: '#/components/schemas/models.Meta'
                  status:
                    example: success
                    type: string
                type: object
          description: Payment methods retrieved successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid shop ID
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Shop not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Failed to get payment methods
      security:
      - OAuth2AccessCode: []
      summary: Get all payment methods for a shop
      tags:
      - payment-methods
  /shops/{shop_id}/payment-methods/{method_type}:
    delete:
      description: Remove a payment method configuration from a shop
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      - description: Payment method type (stripe, paypal, paystack, flutterwave)
        in: path
        name: method_type
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.SuccessResponse'
          description: Payment method deleted successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid request body or shop ID
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Failed to delete payment method
      security:
      - OAuth2AccessCode: []
      summary: Delete a payment method
      tags:
      - payment-methods
    put:
      description: Create or update a payment method configuration for a shop
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      - description: Payment method type (stripe, paypal, paystack, flutterwave)
        in: path
        name: method_type
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.PaymentMethodConfig'
        description: Payment method configuration
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                description: Generic API response
                properties:
                  code:
                    example: 200
                    type: integer
                  data: {}
                  message:
                    example: Object updated successfully
                    type: string
                  meta:
                    $ref: '#/components/schemas/models.Meta'
                  status:
                    example: success
                    type: string
                type: object
          description: Payment method updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid request body or shop ID
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Shop not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Failed to update payment method
      security:
      - OAuth2AccessCode: []
      summary: Create or update a payment method
      tags:
      - payment-methods
  /shops/{shop_id}/payment-methods/{method_type}/status:
    patch:
      description: Enable or disable a payment method for a shop
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      - description: Payment method type (stripe, paypal, paystack, flutterwave)
        in: path
        name: method_type
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              properties:
                is_enabled:
                  type: boolean
              type: object
        description: Payment method status
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                description: Generic API response
                properties:
                  code:
                    example: 200
                    type: integer
                  data: {}
                  message:
                    example: Object updated successfully
                    type: string
                  meta:
                    $ref: '#/components/schemas/models.Meta'
                  status:
                    example: success
                    type: string
                type: object
          description: Payment method status updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid request body or shop ID
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Payment method not found
      security:
      - OAuth2AccessCode: []
      summary: Update payment method status
      tags:
      - payment-methods
  /shops/{shop_id}/payment-methods/{method_type}/test:
    post:
      description: Test the connectivity and configuration of a payment method
      parameters:
      - description: Shop ID
        in: path
        name: shop_id
        required: true
        schema:
          type: string
      - description: Payment method type (stripe, paypal, paystack, flutterwave)
        in: path
        name: method_type
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                description: Generic API response
                properties:
                  code:
                    example: 200
                    type: integer
                  data: {}
                  message:
                    example: Object updated successfully
                    type: string
                  meta:
                    $ref: '#/components/schemas/models.Meta'
                  status:
                    example: success
                    type: string
                type: object
          description: Payment method test completed
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid request or testing not supported
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Payment method not configured
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.ErrorResponse'
          description: Invalid payment method configuration
      security:
      - OAuth2AccessCode: []
      summary: Test payment method configuration
      tags:
      - payment-methods
...
servers:
- url: http://127.0.0.1:8080/v1